+++ src/configuration/stm32Config.ts
@@ -56,11 +56,17 @@ floatAbi: ${config.floatAbi}
 ldscript: ${config.ldscript} # linker script
 
 # Compiler definitions. The -D prefix for the compiler will be automatically added.
-# A path to files with a definition per line in them can be included here.
-# For these files it is assumed that they are prefixed with a dot e.g. a ".definitions".
 cDefinitions: ${createYamlArray(config.cDefinitions)}
 cxxDefinitions: ${createYamlArray(config.cxxDefinitions)}
 asDefinitions: ${createYamlArray(config.asDefinitions)}
+
+# Compiler definition files. you can add a single files or an array of files for different definitions.
+# The file is expected to have a definition each new line.
+# This allows to include for example a .definition file which can be ignored in git and can contain
+# This can be convenient for passing along secrets at compile time, or generating a file for per device setup.
+cDefinitionsFile:
+cxxDefinitionsFile:
+asDefinitionsFile:
 
 # Compiler flags
 cFlags: ${createYamlArray(config.cFlags)}