+++ src/getInfo/getDotDefinitions.ts
@@ -2,20 +2,6 @@ import { workspace, Uri, window } from 'vscode';
 import * as stripComments from 'strip-comments';
 import { join } from 'path';
 /**
- * Separates definitions from files. If a define comes with a . at the beginning it is considered as a file
- * @param definitions unfiltered list of includes
- * @returns a filterd list of includes and .include files
- */
-export function filterDotDefinitionFiles(definitions: string[]): { definitions: string[], dotDefinitions: string[] } {
-  const filteredDefinitions = definitions.filter((def) => (def.charAt(0) !== '.'));
-  const dotDefinitions = definitions.filter((def) => (def.charAt(0) === '.'));
-  return {
-    definitions: filteredDefinitions,
-    dotDefinitions,
-  };
-}
-
-/**
  * Gets NAME=VALUE pairs from a file and ignores comments.
  * @param file input include file, which is a DEFINTION=VALUE file separated by new lines. Can have # comments
  * @returns the include name value pairs in an array
@@ -32,26 +18,25 @@ export function getDefinitionsFromFile(file: string): string[] {
   return resultingList;
 }
 
-/**
- * Gets the definitions and definitions from the files from the files indicated by a predicating dot
- * @param definitions list of definitions, which can contain files which are indicated by a preceding dot.
- */
-export async function getDefinitionsandDefinitionsFromFile(
-  definitions: string[],
-  workspaceLocation: string
+export default async function getDefinitionsFromFiles(
+  workspaceLocation: string, files?: string[] | string
 ): Promise<string[]> {
-  const filteredDefinitions = filterDotDefinitionFiles(definitions);
-  const definitionFilePromises = filteredDefinitions.dotDefinitions.map((fileName) => {
-    const absoluteFilePath = join(workspaceLocation, fileName);
+  let output: string[] = [];
+  if (!files) {
+    return output;
+  }
+  const fileArray = typeof files === 'string' ? [files] : files;
+  const filePromises = fileArray.map((file) => {
+    const absoluteFilePath = join(workspaceLocation, file);
     return workspace.fs.readFile(Uri.file(absoluteFilePath));
   });
   try {
-    const definitionFiles = await Promise.all(definitionFilePromises);
+    const definitionFiles = await Promise.all(filePromises);
     const includesFromFiles = definitionFiles.map((file) => {
       return getDefinitionsFromFile(Buffer.from(file).toString('utf8'));
     });
     includesFromFiles.forEach((defs) => {
-      filteredDefinitions.definitions = filteredDefinitions.definitions.concat(defs);
+      output = output.concat(defs);
     });
 
   } catch (err) {
@@ -59,5 +44,8 @@ export async function getDefinitionsandDefinitionsFromFile(
       `An error occured while reading the provided .include files, please make sure the file is present. Error: ${err}`
     );
   }
-  return filteredDefinitions.definitions;
+  return output;
+
 }
+
+